some progress

+++ includes/cube3d.h
@@ -6,7 +6,7 @@
 /*   By: tpereira <tpereira@42Lisboa.com>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/28 16:50:38 by jibanez-          #+#    #+#             */
-/*   Updated: 2023/03/15 12:05:10 by tpereira         ###   ########.fr       */
+/*   Updated: 2023/03/15 19:56:31 by tpereira         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -183,7 +183,7 @@ typedef struct s_raycast
 	float	dist;
 	bool	hit;
 	int		color;
-	float	offset;
+	int		offset;
 	t_img	texture;
 }				t_raycast;
